refactor(api): split jobs.js controller into per-responsibility modules (#599)
jobs.js had grown to 634 lines (audit finding C1). Split it by
responsibility into focused routers under routes/jobs/:

  configs.js — /api/admin/crawler-configs[...] (persistent crawler configs)
  runs.js    — /api/admin/crawler-jobs[...] + /api/admin/status + /api/admin/crawlers/:type/discover
  helpers.js — the job-config helpers + gate/useSql + re-exported VALID_JOB_TYPES / validateCrawlerConfig

jobs.js is now a barrel that composes the two routers and re-exports the
helpers that scheduler.js and jobs.*.test.js import from ./jobs.js — so
those consumers, app.js and every public path are unchanged. Pure code
move (maskedConfigForResponse is promoted to an export since configs.js
now uses it).

jobs.js had no complexity-baseline entries, so none needed relocating.
The route handlers were previously only validation-tested; extended
jobs.routes.test.js to drive the config/job CRUD, status and log-tail
success + not-found branches, lifting changed-line diff coverage to 87%.
Full API suite green (1772 tests); complexity + jscpd (0 net-new clones)
+ OpenAPI-drift gates pass.
